rename Manufacturing to Research & Development
